# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProjectCRUD",
        "description": "API documentation for the ProjectCRUD application.",
        "version": "1.0.0"
    },
    "paths": {
        "/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get list of cars",
                "description": "Display a listing of the resource.",
                "operationId": "indexCars",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Store a newly created car in storage",
                "description": "Store a newly created resource in storage.",
                "operationId": "storeCar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpsertCarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Successful operation - Redirect to index",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cars/create": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Show the form for creating a new car",
                "description": "Show the form for creating a new resource.",
                "operationId": "createCar",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "carmake": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "enum": [
                                                    "Audi",
                                                    "BMW",
                                                    "Ford",
                                                    "Jaguar",
                                                    "Kia",
                                                    "Land Rover",
                                                    "Lexus",
                                                    "Mercedes Benz",
                                                    "MINI",
                                                    "Nissan",
                                                    "SEAT",
                                                    "Skoda",
                                                    "Toyota",
                                                    "Vauxhall",
                                                    "Volkswagen",
                                                    "Volvo"
                                                ]
                                            }
                                        },
                                        "carbodyshape": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "enum": [
                                                    "Combi",
                                                    "Covertible",
                                                    "Coupe",
                                                    "Estate",
                                                    "Hatchback",
                                                    "MPV",
                                                    "Saloon",
                                                    "Sedan",
                                                    "SUV"
                                                ]
                                            }
                                        },
                                        "carfuel": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "enum": [
                                                    "Electric",
                                                    "Hybrid",
                                                    "Petrol",
                                                    "Oil"
                                                ]
                                            }
                                        },
                                        "cartransmission": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "enum": [
                                                    "Manual",
                                                    "Automatic"
                                                ]
                                            }
                                        },
                                        "carstatus": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "enum": [
                                                    "Fabric",
                                                    "Modified"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cars/{car}/download": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Download image of the specified car from storage",
                "description": "Download image of the specified resource in storage.",
                "operationId": "downloadCarImage",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID of the car to download image from",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation - Download the car image",
                        "content": {
                            "image/jpeg": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpsertCarRequest": {
                "title": "UpsertCarRequest",
                "description": "Request data for creating/updating a car.",
                "required": [
                    "make",
                    "model",
                    "body_shape",
                    "fuel",
                    "transmission",
                    "engine",
                    "engine_capacity",
                    "status",
                    "day_repayment",
                    "rental_date",
                    "return_date",
                    "image"
                ],
                "properties": {
                    "make": {
                        "description": "The make of the car.",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "description": "The model of the car.",
                        "type": "string",
                        "example": "Camry"
                    },
                    "body_shape": {
                        "description": "The body shape of the car.",
                        "type": "string",
                        "example": "Sedan"
                    },
                    "fuel": {
                        "description": "The type of fuel used by the car.",
                        "type": "string",
                        "example": "Petrol"
                    },
                    "transmission": {
                        "description": "The transmission type of the car.",
                        "type": "string",
                        "example": "Automatic"
                    },
                    "engine": {
                        "description": "The engine type of the car.",
                        "type": "string",
                        "example": "V6"
                    },
                    "engine_capacity": {
                        "description": "The engine capacity of the car.",
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "status": {
                        "description": "The status of the car.",
                        "type": "string",
                        "example": "Fabric"
                    },
                    "day_repayment": {
                        "description": "The day repayment of the car.",
                        "type": "number",
                        "example": 10
                    },
                    "rental_date": {
                        "description": "The rental date of the car.",
                        "type": "string",
                        "format": "date",
                        "example": "2023-07-19"
                    },
                    "return_date": {
                        "description": "The return date of the car.",
                        "type": "string",
                        "format": "date",
                        "example": "2023-07-25"
                    },
                    "image": {
                        "description": "The image of the car.",
                        "type": "string",
                        "example": "car_image.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    }
}